+++ src/pages/CurrentFishing.tsx
@@ -1,41 +1,37 @@
-import { useContext, useEffect, useState } from 'react';
+import { useContext, useEffect } from 'react';
 import { useQuery } from 'react-query';
-import { useNavigate } from 'react-router-dom';
-import Button, { ButtonColors } from '../components/buttons/Button';
 import FishingActions from '../components/containers/FishingActions';
 import FishingLocation from '../components/containers/FishingLocation';
 import FishingTools from '../components/containers/FishingTools';
 import FishingToolWeight from '../components/containers/FishingToolWeight';
 import FishingWeight from '../components/containers/FishingWeight';
 import DefaultLayout from '../components/layouts/DefaultLayout';
-import PopUpWithImage from '../components/layouts/PopUpWithImage';
-import { Grid } from '../components/other/CommonStyles';
-import { IconName } from '../components/other/Icon';
 import LoaderComponent from '../components/other/LoaderComponent';
 import {
   getCurrentRoute,
   handleErrorToast,
   LOCATION_ERRORS,
+  PopupContentType,
   slugs,
   validationTexts,
 } from '../utils';
 import api from '../utils/api';
-import { LocationContext, LocationContextType } from '../components/other/LocationContext';
-
-const infoToEnableTheLocationUrl = 'https://zuvys.biip.lt/dokumentacija/zvejyba/lokacija/';
+import { LocationContext, LocationContextType } from '../components/providers/LocationContext';
+import { PopupContext, PopupContextProps } from '../components/providers/PopupProvider';
 
 export const CurrentFishing = () => {
+  const { showPopup } = useContext<PopupContextProps>(PopupContext);
   const { coordinates, error } = useContext<LocationContextType>(LocationContext);
-  const navigate = useNavigate();
   const currentRoute = getCurrentRoute(window.location.pathname);
-  const [showLocationDeniedPopUp, setShowLocationDeniedPopUp] = useState(false);
 
   useEffect(() => {
     if (error === LOCATION_ERRORS.POSITION_UNAVAILABLE) {
       handleErrorToast(validationTexts.failedToSetLocation);
     }
-
-    setShowLocationDeniedPopUp(error === LOCATION_ERRORS.POINT_NOT_FOUND);
+    showPopup({
+      type: PopupContentType.LOCATION_PERMISSION,
+      content: {},
+    });
   }, [error]);
 
   const { data: currentFishing, isLoading: currentFishingLoading } = useQuery(
@@ -46,16 +42,6 @@ export const CurrentFishing = () => {
     },
   );
 
-  useEffect(() => {
-    if (!currentFishingLoading) {
-      if (currentFishing && currentRoute?.slug === slugs.fishingLocation) {
-        navigate(slugs.fishingCurrent);
-      } else if (!currentFishing && currentRoute?.slug !== slugs.fishingLocation) {
-        navigate(slugs.fishingLocation);
-      }
-    }
-  }, [currentFishing, window.location.pathname]);
-
   if (currentFishingLoading) {
     return <LoaderComponent />;
   }
@@ -64,48 +50,25 @@ export const CurrentFishing = () => {
 
   return (
     <DefaultLayout>
-      <PopUpWithImage
-        onClose={() => setShowLocationDeniedPopUp(false)}
-        iconName={IconName.location}
-        visible={showLocationDeniedPopUp}
-        title={'Geografinės vietos leidimas'}
-        description={
-          'Norint naudotis elektroniniu žvejybos žurnalu, prašome suteikti leidimą naudotis jūsų geografinę vietą.'
-        }
-      >
-        <Grid>
-          <Button
-            onClick={() => {
-              window.location.href = infoToEnableTheLocationUrl;
-            }}
-          >
-            {'Skaityti instrukciją'}
-          </Button>
-          <Button
-            variant={ButtonColors.SECONDARY}
-            onClick={() => setShowLocationDeniedPopUp(false)}
-          >
-            {'Žinau kaip suteikti leidimą'}
-          </Button>
-        </Grid>
-      </PopUpWithImage>
-
-      {currentRoute?.slug === slugs.fishingLocation && (
+      {currentFishing?.id ? (
         <FishingLocation isDisabled={isDisabled} coordinates={coordinates} />
-      )}
-      {currentRoute?.slug === slugs.fishingCurrent && (
-        <FishingActions
-          isDisabled={isDisabled}
-          fishing={currentFishing}
-          coordinates={coordinates}
-        />
-      )}
-      {currentRoute?.slug === slugs.fishingTools && <FishingTools />}
-      {currentRoute?.slug === slugs.fishingToolCaughtFishes(':toolId') && (
-        <FishingToolWeight isDisabled={isDisabled} coordinates={coordinates} />
-      )}
-      {currentRoute?.slug === slugs.fishingWeight && (
-        <FishingWeight isDisabled={isDisabled} coordinates={coordinates} />
+      ) : (
+        <>
+          {currentRoute?.slug === slugs.fishingCurrent && (
+            <FishingActions
+              isDisabled={isDisabled}
+              fishing={currentFishing}
+              coordinates={coordinates}
+            />
+          )}
+          {currentRoute?.slug === slugs.fishingTools && <FishingTools />}
+          {currentRoute?.slug === slugs.fishingToolCaughtFishes(':toolId') && (
+            <FishingToolWeight isDisabled={isDisabled} coordinates={coordinates} />
+          )}
+          {currentRoute?.slug === slugs.fishingWeight && (
+            <FishingWeight isDisabled={isDisabled} coordinates={coordinates} />
+          )}
+        </>
       )}
     </DefaultLayout>
   );
